Update with KES changes in cardano-base.
- Use SUM KES scheme
- KES updating must now take the current period, rather than the target
period.
- The KES class now no longer tracks its current period. As such, we
introduce an abstraction `HotKey` which collects the current period with
the KES key.

+++ ouroboros-consensus/ouroboros-consensus-mock/src/Ouroboros/Consensus/Mock/Node/Praos.hs
@@ -54,13 +54,13 @@ protocolInfoPraos numCoreNodes nid params eraParams =
       }
   where
     signKeyVRF :: CoreNodeId -> SignKeyVRF MockVRF
-    signKeyVRF (CoreNodeId n) = SignKeyMockVRF (fromIntegral n)
+    signKeyVRF (CoreNodeId n) = SignKeyMockVRF n
 
     verKeyVRF :: CoreNodeId -> VerKeyVRF MockVRF
-    verKeyVRF (CoreNodeId n) = VerKeyMockVRF (fromIntegral n)
+    verKeyVRF (CoreNodeId n) = VerKeyMockVRF n
 
     verKeyKES :: CoreNodeId -> VerKeyKES (MockKES t)
-    verKeyKES (CoreNodeId n) = VerKeyMockKES (fromIntegral n)
+    verKeyKES (CoreNodeId n) = VerKeyMockKES n
 
     addrDist :: AddrDist
     addrDist = mkAddrDist numCoreNodes
